+++ docs/tr1/CHANGELOG.md
@@ -7,6 +7,7 @@
 - added support for key/puzzle/pickup descriptions, allowing players to examine said items in the inventory (#1821)
 - added an option to fix inventory item usage duplication (#1586)
 - added optional automatic key/puzzle inventory item pre-selection (#1884)
+- added a search feature to the config tool (#1889)
 - changed OpenGL backend to use version 3.3, with fallback to 2.1 if initialization fails (#1738)
 - changed text backend to accept named sequences. Currently supported sequences (limited by the sprites available in OG):
     - `\{umlaut}`