+++ docs/tr2/CHANGELOG.md
@@ -2,6 +2,7 @@
 - added support for custom levels to enforce values for any config setting (#1846)
 - added an option to fix inventory item usage duplication (#1586)
 - added optional automatic key/puzzle inventory item pre-selection (#1884)
+- added a search feature to the config tool (#1889)
 - fixed depth problems when drawing certain rooms (#1853, regression from 0.6)
 - fixed Lara getting stuck in her hit animation if she is hit while mounting the boat or skidoo (#1606)
 - fixed being unable to go from surface swimming to underwater swimming without first stopping (#1863, regression from 0.6)